Move wallet info into table
